+++ react/components/utils/index.js
@@ -7,3 +7,5 @@ export const childrenOf = (...types) => {
 
   return PropTypes.oneOfType([fieldType, PropTypes.arrayOf(fieldType)])
 }
+
+export const isMobileDevice = device => device.toLowerCase() === 'phone'